Reformatted project
