+++ src/App.css
@@ -1,4 +1,4 @@
 @import "tailwindcss";
-@plugin "./hero.ts";
+@plugin "./helper/hero.ts";
 @source '../node_modules/@heroui/theme/dist/**/*.{js,ts,jsx,tsx}';
 @custom-variant dark (&:is(.dark *));
