+++ src/helper/categories.tsx
@@ -1,0 +1,11 @@
+const categories = [
+  { value: "P&A", label: "Pops and Anime" },
+  { value: "NN", label: "niconico" },
+  { value: "TP", label: "Touhou Project" },
+  { value: "Variety", label: "Variety" },
+  { value: "GekiMai", label: "GekiMai" },
+  { value: "Irodorimidori", label: "Irodorimidori" },
+  { value: "Original", label: "Original" },
+];
+
+export default categories;
