+++ src/helper/versions.tsx
@@ -1,0 +1,23 @@
+const versions = [
+  { value: "Chunithm", label: "Chunithm" },
+  { value: "Chunithm Plus", label: "Chunithm Plus" },
+  { value: "Air", label: "Air" },
+  { value: "Air Plus", label: "Air Plus" },
+  { value: "Star", label: "Star" },
+  { value: "Star Plus", label: "Star Plus" },
+  { value: "Amazon", label: "Amazon" },
+  { value: "Amazon Plus", label: "Amazon Plus" },
+  { value: "Crystal", label: "Crystal" },
+  { value: "Crystal Plus", label: "Crystal Plus" },
+  { value: "Paradise", label: "Paradise" },
+  { value: "Paradise Lost", label: "Paradise Lost" },
+  { value: "New", label: "New" },
+  { value: "New Plus", label: "New Plus" },
+  { value: "Sun", label: "Sun" },
+  { value: "Sun Plus", label: "Sun Plus" },
+  { value: "Luminous", label: "Luminous" },
+  { value: "Luminous Plus", label: "Luminous Plus" },
+  { value: "Verse", label: "Verse" },
+];
+
+export default versions;
